Optimized # print Sub strings
The code provided is already optimized, but here are a few additional suggestions:

Add a try-with-resources statement to close the Scanner object automatically when the input is complete.

Rename the print_sub_String method to printSubstrings to follow the standard Java naming conventions for methods.

Use System.lineSeparator() to add a newline instead of "\n" to make the code more platform-independent.

------

which one is better approach :

1: try-with-resources statement to close the Scanner object

2: scanner.close()

------------

Using a try-with-resources statement to close the Scanner object is generally considered to be a better approach than manually calling scanner.close().

The main advantage of using a try-with-resources statement is that it ensures that the Scanner object is always closed, even if an exception is thrown. This is because the try-with-resources statement automatically calls the close() method on the Scanner object when the try block is exited, whether normally or due to an exception. This helps to prevent resource leaks and makes the code more robust.

In contrast, manually calling scanner.close() can be error-prone, as it is possible to forget to call it, or to call it in the wrong place. If the close() method is not called, the resource (in this case, the Scanner object) will remain open and may cause issues in the program.

So, in summary, using a try-with-resources statement to close the Scanner object is generally the better approach, as it ensures that the Scanner object is always properly closed and helps to prevent resource leaks.